reperto-backend/requirements.txt
fastapi
uvicorn
pydantic
python-multipart
sqlalchemy
databases
passlib[bcrypt]
pyjwt
openai
python-dotenv